+++ recipes/views.py
@@ -7,8 +7,14 @@ from recipes.models import Recipe
 
 def catalogue(request):
     all_recipes = Recipe.objects.all()
+    profile = Profile.objects.first()
+
+    if not profile:  # Plays the role of a guard so that noone can manually type the path in the url
+        return redirect('home')
+
     context = {
-        "all_recipes": all_recipes
+        "all_recipes": all_recipes,
+        "profile": profile
     }
     return render(request, 'recipes/catalogue.html', context)
 
@@ -16,6 +22,9 @@ def catalogue(request):
 def create_recipe(request):
     form = RecipeCreateForm(request.POST or None)
     profile = Profile.objects.first()
+
+    if not profile:  # Plays the role of a guard so that noone can manually type the path in the url
+        return redirect('home')
 
     if request.method == 'POST':
         if form.is_valid():
@@ -25,7 +34,8 @@ def create_recipe(request):
             return redirect('catalogue')
 
     context = {
-        "form": form
+        "form": form,
+        "profile": profile
     }
 
     return render(request, 'recipes/create-recipe.html', context)
@@ -33,9 +43,16 @@ def create_recipe(request):
 
 def details_recipe(request, recipe_id):
     recipe = Recipe.objects.get(id=recipe_id)
+    profile = Profile.objects.first()
+
+
+    if not profile:  # Plays the role of a guard so that noone can manually type the path in the url
+        return redirect('home')
 
     context = {
-        "recipe": recipe
+        "recipe": recipe,
+        "profile": profile
+
     }
     return render(request, 'recipes/details-recipe.html', context)
 
@@ -43,6 +60,10 @@ def details_recipe(request, recipe_id):
 def edit_recipe(request, recipe_id):
     recipe = Recipe.objects.get(id=recipe_id)
     form = RecipeEditForm(request.POST or None, instance=recipe)
+    profile = Profile.objects.first()
+
+    if not profile:  # Plays the role of a guard so that noone can manually type the path in the url
+        return redirect('home')
 
     if request.method == 'POST':
         if form.is_valid():
@@ -50,7 +71,8 @@ def edit_recipe(request, recipe_id):
             return redirect('catalogue')
 
     context = {
-        "form": form
+        "form": form,
+        "profile": profile
     }
     return render(request, 'recipes/edit-recipe.html', context)
 
@@ -58,13 +80,18 @@ def edit_recipe(request, recipe_id):
 def delete_recipe(request, recipe_id):
     recipe = Recipe.objects.get(id=recipe_id)
     form = RecipeDeleteForm(request.POST or None, instance=recipe)
+    profile = Profile.objects.first()
+
+    if not profile:  # Plays the role of a guard so that noone can manually type the path in the url
+        return redirect('home')
 
     if request.method == 'POST':
         recipe.delete()
         return redirect('catalogue')
 
     context = {
-        "form": form
+        "form": form,
+        "profile": profile
     }
 
     return render(request, 'recipes/delete-recipe.html', context)
